Moved push files to dedicated directory.
