+++ src/App.tsx
@@ -20,7 +20,7 @@ import { Web3Modal } from '@web3modal/react';
 
 import { chain, configureChains, createClient, WagmiConfig } from 'wagmi';
 import Messaging from './pages/Messaging';
-import { PushProvider } from './messaging/context/pushUser';
+import { PushProvider } from './messaging/push/context/pushUser';
 
 const chains = [chain.goerli];
