+++ src/messaging/push/components/ConversationCard.tsx
@@ -1,11 +1,11 @@
 import { shortAddress, truncate } from '../utils/messaging';
-import useUserByAddress from '../../hooks/useUserByAddress';
+import useUserByAddress from '../../../hooks/useUserByAddress';
 import { useNavigate } from 'react-router-dom';
-import { ConversationDisplayType } from '../../types';
+import { ConversationDisplayType } from '../../../types';
 import { chat as chatApi } from '@pushprotocol/restapi/src/lib';
 import { pCAIP10ToWallet } from '@pushprotocol/restapi/src/lib/helpers';
 import { CheckCircle } from 'heroicons-react';
-import { formatDateConversationCard } from '../../utils/dates';
+import { formatDateConversationCard } from '../../../utils/dates';
 
 interface IConversationCardProps {
   address: string;
