+++ src/messaging/push/components/MessageComposer.tsx
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction } from 'react';
-import Loading from '../../components/Loading';
+import Loading from '../../../components/Loading';
 
 interface IMessageComposerProps {
   messageContent: string;
