+++ src/messaging/push/utils/messaging.ts
@@ -1,6 +1,6 @@
 import { Message } from '@pushprotocol/restapi/src/lib/chat/ipfs';
 import { IMessageIPFS } from '@pushprotocol/uiweb/lib/types';
-import { ChatMessage, ChatMessageStatus, PushMessage } from '../../types';
+import { ChatMessage, ChatMessageStatus, PushMessage } from '../../../types';
 
 export const shortAddress = (addr: string) =>
   addr.length > 10 && addr.startsWith('0x')
